test insomnia.sendRequest
